[ActivityModeration] Handling of the delete form in NoticeOptionItems.
The majority of the changes concerns with making the plugin control
both the UI and logic aspects of the delete form in NoticeOptionItems.
Other changes are minor fixes or updates to the code visited during
the work.

ActivityModeration isn't still converted to a full "ActivityVerbHandlerPlugin",
that will require some time to properly do due to the changes done
to the core delete logic when this plugin was first introduced.